Logout of Server while Editing
The system log's out the server while user performs editing. This is because the server times out and doesn't inform the user. This can later cause sudden failures during the code upload phase. So we logout, save credentials, and then reconnect to server while uploading. Credential verification is already performed before this stage.